+++ src/main/java/ru/gadjini/reminder/bot/command/state/ReminderRequestData.java
@@ -9,9 +9,9 @@ public class ReminderRequestData {
 
     private String receiverName;
 
-    private Integer receiverId;
+    private Long receiverId;
 
-    private Integer creatorId;
+    private Long creatorId;
 
     private String text;
 
@@ -23,11 +23,11 @@ public class ReminderRequestData {
 
     private String language;
 
-    public Integer getReceiverId() {
+    public Long getReceiverId() {
         return receiverId;
     }
 
-    public void setReceiverId(Integer receiverId) {
+    public void setReceiverId(Long receiverId) {
         this.receiverId = receiverId;
     }
 
@@ -67,11 +67,11 @@ public class ReminderRequestData {
         this.messageId = messageId;
     }
 
-    public Integer getCreatorId() {
+    public Long getCreatorId() {
         return creatorId;
     }
 
-    public void setCreatorId(Integer creatorId) {
+    public void setCreatorId(Long creatorId) {
         this.creatorId = creatorId;
     }
