+++ src/main/java/ru/gadjini/reminder/dao/paymentmessage/PaymentMessageDao.java
@@ -14,7 +14,7 @@ public class PaymentMessageDao {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    public Integer getMessageId(int userId) {
+    public Integer getMessageId(long userId) {
         return jdbcTemplate.query(
                 "SELECT message_id FROM payment_message WHERE user_id = ?",
                 ps -> ps.setLong(1, userId),
@@ -28,14 +28,14 @@ public class PaymentMessageDao {
         );
     }
 
-    public void create(int userId, int messageId) {
+    public void create(long userId, int messageId) {
         jdbcTemplate.update(
                 "INSERT INTO payment_message(user_id, message_id) VALUES (" + userId + ", " + messageId + ") " +
                         "ON CONFLICT (user_id) DO UPDATE SET message_id = excluded.message_id"
         );
     }
 
-    public Integer delete(int userId) {
+    public Integer delete(long userId) {
         return jdbcTemplate.query(
                 "WITH pm AS (DELETE FROM payment_message WHERE user_id = ? RETURNING message_id) SELECT * FROM pm",
                 ps -> ps.setLong(1, userId),
